refactor: decompose lang component in the chat component, create LangSelect component
